requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
cssselect2==0.7.0
Django==4.2
django-admin-sortable2==2.2.1
django-compat==1.0.15
django-filer==3.1.3
django-polymorphic==3.1.0
django-reversion==5.0.12
django_threaded_sync_to_async==1.0.4
easy-thumbnails==2.8.5
frozenlist==1.4.1
google-auth==2.31.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
idna==3.7
lxml==5.2.2
magic-filter==1.0.12
multidict==6.0.5
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyMySQL==1.1.1
pyTelegramBotAPI==4.20.0
pytesseract==0.3.10
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
webencodings==0.5.1
yarl==1.9.4
